requirements.txt
astroid==2.2.5
autopep8==1.4.3
certifi==2019.3.9
chardet==3.0.4
Click==7.0
Django==2.1.7
django-crispy-forms==1.7.2
django-filter==2.1.0
django-progressbarupload==0.1.7
django-simple-history==2.7.0
djangorestframework==3.9.2
idna==2.8
isort==4.3.13
lazy-object-proxy==1.3.1
Markdown==3.0.1
mccabe==0.6.1
numpy==1.16.2
opencv-contrib-python==3.4.5.20
pep8==1.7.1
Pillow==5.4.1
psycopg2==2.7.7
psycopg2-binary==2.7.7
pycodestyle==2.5.0
pylint==2.3.1
pytz==2018.9
rarfile==3.0
requests==2.21.0
six==1.12.0
torch==1.0.1.post2
torchtext==0.3.1
torchvision==0.2.2.post3
tqdm==4.31.1
typed-ast==1.3.1
unrar==0.3
urllib3==1.24.1
whitenoise==4.1.2
wrapt==1.11.1
